requirements.txt
scrapy==2.8.0
psycopg2-binary==2.9.5
python-dotenv=0.21.1